+++ src/main/kotlin/no/ok/origo/dataplatform/commons/auth/AuthToken.kt
@@ -1,9 +1,9 @@
 package no.ok.origo.dataplatform.commons.auth
 import com.fasterxml.jackson.annotation.JsonProperty
-import com.fasterxml.jackson.databind.PropertyNamingStrategy
+import com.fasterxml.jackson.databind.PropertyNamingStrategies
 import com.fasterxml.jackson.databind.annotation.JsonNaming
 
-@JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy::class)
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
 data class AuthToken(
     val accessToken: String,
     val expiresIn: Int,